✨Ref: free_board 리팩토링

+++ app/domain/free_board/services.py
@@ -7,7 +7,7 @@ from app.domain.free_board.repository import (
     get_free_boards_query,
     patch_free_board_by_id,
 )
-from app.domain.free_board.validator import author_board
+from app.utils.auth import check_author
 from app.utils.exception import check_existing
 
 
@@ -31,7 +31,7 @@ async def patch_free_board_by_id_service(id: int, free_board: Any, current_user:
     """업데이트"""
     board = await get_free_board_query(id)
     check_existing(board, "해당 게시글을 찾을 수 없습니다.", "free_board_not_found")
-    author_board(board, current_user)
+    await check_author(board, current_user)
     board = await patch_free_board_by_id(board, free_board)
 
     return board
@@ -41,5 +41,5 @@ async def delete_free_board_by_id_service(id: int, current_user):
     """삭제"""
     board = await get_free_board_query(id)
     check_existing(board, "해당 게시글을 찾을 수 없습니다.", "free_board_not_found")
-    author_board(board, current_user)
+    await check_author(board, current_user)
     await delete_free_board_by_id(board)
